refactor: move types into own file
